+++ src/components/hero/heroBarLetters/firstName/E.tsx
