+++ src/components/hero/heroBarLetters/firstName/LowerA.tsx
