+++ src/components/hero/heroBarLetters/firstName/LowerD.tsx
