+++ src/components/hero/heroBarLetters/firstName/LowerN.tsx
